server/REST/requirements.txt
Flask
psycopg2-binary
websockets